Keep validation rules in separate files
