+++ server/src/validationRules/index.ts
@@ -1,0 +1,11 @@
+import colorsThresholds from "./colorsThresholds";
+import forecastEndTime from "./forecastEndTime";
+import forecastSSA from "./forecastSSA";
+import startEndTime from "./startEndTime";
+
+export default [
+    colorsThresholds,
+    forecastEndTime,
+    forecastSSA,
+    startEndTime
+];
